+++ src/main/java/priv/klochkov/island/model/animal/herbivores/Herbivore.java
@@ -4,7 +4,7 @@ import priv.klochkov.island.model.animal.Animal;
 import priv.klochkov.island.model.animal.interfaces.IEatablePlant;
 
 public abstract class Herbivore extends Animal implements IEatablePlant {
-    public Herbivore(int speedMovement, int weight, int maxSatiety) {
+    public Herbivore(int speedMovement, float weight, float maxSatiety) {
         super(speedMovement, weight, maxSatiety);
     }
 }
