+++ src/main/java/priv/klochkov/island/model/animal/predators/Predator.java
@@ -4,7 +4,7 @@ import priv.klochkov.island.model.animal.Animal;
 import priv.klochkov.island.model.animal.interfaces.IEatableAnimal;
 
 public abstract class Predator extends Animal implements IEatableAnimal {
-    public Predator(int speedMovement, int weight, int maxSatiety) {
+    public Predator(int speedMovement, float weight, float maxSatiety) {
         super(speedMovement, weight, maxSatiety);
     }
 }
